+++ plugins/event_notification/providers/kafka/lib/model/KafkaNotificationTemplate.php
@@ -16,6 +16,7 @@ class KafkaNotificationTemplate extends EventNotificationTemplate
 	const CUSTOM_DATA_PARTITION_KEY = 'partitionKey';
 	const CUSTOM_DATA_MESSAGE_FORMAT = 'messageFormat';
 	const CUSTOM_DATA_API_OBJECT_TYPE = 'apiObjectType';
+	const CUSTOM_DATA_RESPONSE_PROFILE_SYSTEM_NAME = 'responseProfileSystemName';
 	const SCHEMA_ID = 'schemaId';
 	const SCHEMA = 'schema';
 	
@@ -75,6 +76,16 @@ class KafkaNotificationTemplate extends EventNotificationTemplate
 	{
 		return $this->getFromCustomData(self::CUSTOM_DATA_API_OBJECT_TYPE);
 	}
+
+	public function setResponseProfileSystemName($value)
+	{
+		return $this->putInCustomData(self::CUSTOM_DATA_RESPONSE_PROFILE_SYSTEM_NAME, $value);
+	}
+
+	public function getResponseProfileSystemName()
+	{
+		return $this->getFromCustomData(self::CUSTOM_DATA_RESPONSE_PROFILE_SYSTEM_NAME);
+	}
 	
 	public function dispatch(kScope $scope)
 	{
@@ -119,9 +130,15 @@ class KafkaNotificationTemplate extends EventNotificationTemplate
 			$modifiedColumns = $scope->getEvent()->getModifiedColumns();
 			$modifiedColumns = $this->buildMessageOldValues($modifiedColumns);
 		}
-		
+
+		$responseProfile = null;
+		if($this->getResponseProfileSystemName())
+		{
+			$responseProfile = ResponseProfilePeer::retrieveBySystemName($this->getResponseProfileSystemName());
+		}
+
 		$apiObjectType = $this->getApiObjectType();
-		$apiObject = call_user_func(kCurrentContext::$serializeCallback, $object, $apiObjectType, 1);
+		$apiObject = call_user_func(kCurrentContext::$serializeCallback, $object, $apiObjectType, 1, $responseProfile);
 		$apiObject = json_decode($apiObject, true);
 		
 		$apiObjectAdditionalParams = $this->getContentParameters();